+++ src/core_graphics/geometry/mod.rs
@@ -1,9 +1,11 @@
+mod cgaffinetransform;
 mod cgfloat;
 mod cgpoint;
 mod cgrect;
 mod cgsize;
 mod cgvector;
 
+pub use cgaffinetransform::*;
 pub use cgfloat::*;
 pub use cgpoint::*;
 pub use cgrect::*;